[test]Add test for running simulation
Add class to contain list of rover test data
Add class to contain simulation test data
Add function that can be given an index and will return the data to test
that simulation
add function that will publically test the sim, this function is a
theory. This method was chosen because there is a relatively large
amount of setup that needs to be done for the simulation to be setup and
run and as this involves heavy use of the new keyword its almost
impossible to pass these details into the function via inline data. As
such I have opted to house it in a separate function that will be called
to get the simulation details. Its not the best solution but it works
well enoughh that ill leave it untill I come up with a better way